Complete Two To One (7kyu)
This was my first 7 kyu from scratch in a minute.  I'm preparing
for an Amazon AWS interview and want to brush up on some problem solving
and need to start somewhere easy-ish.

We dusted off some very useful and common methods in this excersize,
notably .split() and .join().  It was important to remember to include
the qoatation marks in there to create a new element for each character
in the array when splitting, and also in .join('') in order to remove
the commas when putting the array back together into one cohesive
string.

fullStr = (Array.from(new Set(fullStr))); was very new to me, and
I believe it is new-ish to JS.  This was a very quick and easy way to
detect and remove all the duplicate characters from the array.