Add Printed Errors, 7kyu
I successfully implemented a nested for loop in this script.  It checks
each element in an array made up of the input letters against every
element in an array of "bad letters" (letters above m, non-inclusive).
It then puts the bad letters into a new array, and returns and answer
based on the array length of the original input (split into an array)
and the new array containing all the bad letters to create an answer
presented in the format of a simple fraction, as a string of course.