Add Replace with alphabet position (6kyu)
I used a nested for loop to iteratate over each letter from the original
argument string, and compare it to each letter in an alphabet array.
Each time there was a match, the value of the index + 1 (position in
alphabet) was added to an answer array.  Then lastly convert that
array back into a string in the correct format and return it.